dom/tests/mochitest/geolocation/test_windowClose.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=493615
-->
<head>
  <title>Test for geolocation in chrome </title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="geolocation_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=493615">Mozilla Bug 493615</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

function done() {
  ok(1, "no crash, so pass.");
  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
force_prompt(true, function() {
  window.open("windowTest.html");
});

</script>
</pre>
</body>
</html>
